refactor: Refactor pvp code
